+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Features:
 
 - Add the `project.description` to the configuration schema.
+- Provide `callAll` on the `FunctionRegistry` and `WorkspaceContext`.
 
 ## v0.21.1 (2026-02-09)